refactor: verificationtoken 패키지 분리
