+++ src/test/java/com/jobnote/domain/user/service/UserServiceTest.java
@@ -2,14 +2,14 @@ package com.jobnote.domain.user.service;
 
 import com.jobnote.ServiceUnitTest;
 import com.jobnote.domain.user.domain.User;
-import com.jobnote.domain.user.domain.VerificationToken;
+import com.jobnote.domain.verificationtoken.domain.VerificationToken;
 import com.jobnote.domain.user.dto.UserAvatarRequest;
 import com.jobnote.domain.user.dto.UserNicknameRequest;
 import com.jobnote.domain.user.dto.UserProfileResponse;
 import com.jobnote.domain.user.dto.UserSignUpRequest;
 import com.jobnote.domain.user.event.SignUpEvent;
 import com.jobnote.domain.user.repository.UserRepository;
-import com.jobnote.domain.user.repository.VerificationTokenRepository;
+import com.jobnote.domain.verificationtoken.repository.VerificationTokenRepository;
 import com.jobnote.global.common.ResponseCode;
 import com.jobnote.global.exception.JobNoteException;
 import org.junit.jupiter.api.DisplayName;
